+++ src/Firebase/V3/Firebase.php
@@ -2,10 +2,11 @@
 
 namespace Firebase\V3;
 
-use Firebase\Auth\Token\Handler;
+use Firebase\Auth\Token\Handler as TokenHandler;
 use Firebase\Database;
 use Firebase\Database\ApiClient;
 use Firebase\Exception\InvalidArgumentException;
+use Firebase\Factory;
 use Firebase\Http\Middleware;
 use Firebase\ServiceAccount;
 use Firebase\V3\Auth\CustomToken;
@@ -34,13 +35,23 @@ class Firebase
      */
     private $database;
 
-    public function __construct(ServiceAccount $serviceAccount, UriInterface $databaseUri)
+    /**
+     * @var TokenHandler
+     */
+    private $tokenHandler;
+
+    public function __construct(ServiceAccount $serviceAccount, UriInterface $databaseUri, TokenHandler $tokenHandler)
     {
         $this->serviceAccount = $serviceAccount;
         $this->databaseUri = $databaseUri;
+        $this->tokenHandler = $tokenHandler;
     }
 
     /**
+     * Creates a new Firebase instance from a service account.
+     *
+     * @deprecated 2.2 will be removed in 3.0, use Firebase\Factory instead
+     *
      * @param mixed $serviceAccount Service Account (ServiceAccount instance, JSON, array, path to JSON file)
      * @param string|UriInterface $databaseUri
      *
@@ -50,20 +61,44 @@ class Firebase
      */
     public static function fromServiceAccount($serviceAccount, $databaseUri = null)
     {
+        trigger_error(
+            'This method is deprecated and will be removed in release 3.0 of this library.'
+            .' Use Firebase\Factory instead.', E_USER_DEPRECATED
+        );
+
         $serviceAccount = ServiceAccount::fromValue($serviceAccount);
 
         $databaseUri = $databaseUri
             ? Psr7\uri_for($databaseUri)
             : new Uri(sprintf('https://%s.firebaseio.com', $serviceAccount->getProjectId()));
 
-        return new self($serviceAccount, $databaseUri);
+        $factory = new Factory();
+        $rc = new \ReflectionObject($factory);
+        $rm = $rc->getMethod('getDefaultTokenHandler');
+        $rm->setAccessible(true);
+
+        $tokenHandler = $rm->invoke($factory, $serviceAccount);
+
+        return new self($serviceAccount, $databaseUri, $tokenHandler);
     }
 
+    /**
+     * Returns a new instance with a changed database URI.
+     *
+     * @param string|UriInterface $databaseUri
+     *
+     * @return Firebase
+     */
     public function withDatabaseUri($databaseUri): Firebase
     {
-        return new self($this->serviceAccount, Psr7\uri_for($databaseUri));
+        return new self($this->serviceAccount, Psr7\uri_for($databaseUri), $this->tokenHandler);
     }
 
+    /**
+     * Returns an instance of the realtime database.
+     *
+     * @return Database
+     */
     public function getDatabase(): Database
     {
         if (!$this->database) {
@@ -73,6 +108,15 @@ class Firebase
         return $this->database;
     }
 
+    /**
+     * Returns a new instance with the permissions
+     * of the user with the given UID and claims.
+     *
+     * @param string $uid
+     * @param array $claims
+     *
+     * @return Firebase
+     */
     public function asUserWithClaims(string $uid, array $claims = []): Firebase
     {
         return $this->withCustomAuth(new CustomToken($uid, $claims));
@@ -85,20 +129,16 @@ class Firebase
      * @see https://firebase.google.com/docs/auth/admin/create-custom-tokens
      * @see https://firebase.google.com/docs/auth/admin/verify-id-tokens
      *
-     * @return Handler
+     * @return TokenHandler
      */
-    public function getTokenHandler(): Handler
+    public function getTokenHandler(): TokenHandler
     {
-        return new Handler(
-            $this->serviceAccount->getProjectId(),
-            $this->serviceAccount->getClientEmail(),
-            $this->serviceAccount->getPrivateKey()
-        );
+        return $this->tokenHandler;
     }
 
     private function withCustomAuth(Auth $override): Firebase
     {
-        $firebase = new self($this->serviceAccount, $this->databaseUri);
+        $firebase = new self($this->serviceAccount, $this->databaseUri, $this->tokenHandler);
         $firebase->database = $this->getDatabase()->withCustomAuth($override);
 
         return $firebase;